refactor: moved charts under molecules; Added stories decorator
