+++ components/index.js
@@ -11,7 +11,6 @@ export * from './meta'
 
 // Components
 export * from './atoms'
-export * from './charts'
 export * from './molecules'
 export * from './organisms'
 export * from './layouts'
