+++ components/molecules/charts/responsiveBar/__tests__/responsiveBar.test.js
