+++ components/molecules/charts/responsiveBar/index.js
