+++ components/molecules/index.js
@@ -6,6 +6,7 @@ export * from './alert'
 export * from './breadcrumb'
 export * from './buttons'
 export * from './card'
+export * from './charts'
 // export * from './carousel'
 export * from './copyright'
 // export * from './csv'
